requirements.txt
Flask
schedule
pywebpush
python-dotenv
openai
gunicorn
